Navigation functionality of sidebar is created.
